Fixed All Bugs

+++ src/Services/wallet.service.ts
@@ -37,7 +37,7 @@ export class WalletService{
         }
     }
     async AddOne(foreign_id:number,roleid:number,amount:number):Promise<Result<Wallet>>{
-        await this.db.exec(`INSERT INTO wallets (foreign_id,amount) VALUES (${foreign_id},${amount});`);
+        await this.db.exec(`INSERT INTO wallets (foreign_id,roleid,amount) VALUES (${foreign_id},${roleid},${amount});`);
         let wallet:Result<Wallet>=await this.FindOneForeignID(foreign_id,roleid);
         if(wallet.success){
             return new Promise(e=>e({success:true,value:wallet.value} as Result<Wallet>));
@@ -48,7 +48,7 @@ export class WalletService{
     async DeleteOne(foreign_id:number,roleid:number):Promise<Result<Wallet>>{
         let wallet:Result<Wallet>=await this.FindOneForeignID(foreign_id,roleid);
         if(JSON.stringify(wallet.value)!="{}"){
-            await this.db.exec(`DELETE FROM wallets WHERE foreign_id=${foreign_id}`);
+            await this.db.exec(`DELETE FROM wallets WHERE foreign_id=${foreign_id} AND roleid=${roleid}`);
             return new Promise(e=>e({success:true,value:wallet.value} as Result<Wallet>));
         }
         else{
@@ -58,7 +58,7 @@ export class WalletService{
     async UpdateOne(foreign_id:number,roleid:number,amount:number):Promise<Result<Wallet>>{
         let wallet:Result<Wallet>=await this.FindOneForeignID(foreign_id,roleid);
         if(JSON.stringify(wallet.value)!="{}"){
-            await this.db.exec(`UPDATE wallets SET amount=${amount} WHERE foreign_id=${foreign_id};`);
+            await this.db.exec(`UPDATE wallets SET amount=${amount} WHERE foreign_id=${foreign_id} AND roleid=${roleid};`);
             let wallet_updated:Result<Wallet>=await this.FindOneForeignID(foreign_id,roleid);
             return new Promise(e=>e({success:true,value:wallet_updated.value} as Result<Wallet>));
         }else{
